requirements.txt
pyautogen==0.2.27  # 버전 명시 (버전업에 따른 API 변경 방지)
python-dotenv==1.0.1
pandas==2.2.0      # 결과 분석용